backend/requirements.txt
pyramid==2.0
cornice
waitress
transformers
torch
python-dotenv